+++ src/views/account/AccountViewContentTop.vue
@@ -1,20 +1,20 @@
 <template>
   <v-row>
     <v-col cols="12" md="4">
-      <AccountViewAddress />
+      <AccountViewIdentity />
     </v-col>
     <v-col cols="12" md="4">
       <AccountViewChartPerformance :assets="capital.assets" />
     </v-col>
     <v-col cols="12" md="4">
-      <AccountViewBank :balances="balances" />
+      <AccountViewTokensBalance :balances="balances" />
     </v-col>
   </v-row>
 </template>
 
 <script>
-import AccountViewAddress from './AccountViewAddress.vue';
-import AccountViewBank from './AccountViewBank.vue';
+import AccountViewIdentity from './AccountViewIdentity.vue';
+import AccountViewTokensBalance from './AccountViewTokensBalance.vue';
 import AccountViewChartPerformance from './AccountViewChartPerformance.vue';
 
 import accountBalanceHelper from './helpers/accountBalanceHelper';
@@ -23,8 +23,8 @@ import { mapGetters } from 'vuex';
 export default {
   name: 'AccountViewContentTop',
   components: {
-    AccountViewAddress,
-    AccountViewBank,
+    AccountViewIdentity,
+    AccountViewTokensBalance,
     AccountViewChartPerformance,
   },
   computed: {