+++ src/views/validators/detail/ValidatorsViewDetailContentBottom.vue
@@ -1,25 +1,23 @@
 <template>
   <v-row>
     <v-col cols="12" md="4">
-      <ValidatorsViewDetailContentBottomLeft
+      <ValidatorsViewDetailDelegatedTokens
         :delegations="processedDelegations"
       />
     </v-col>
     <v-col cols="12" md="4">
-      <ValidatorsViewDetailContentBottomCenter
-        :delegations="processedDelegations"
-      />
+      <ValidatorsViewDetailChart :delegations="processedDelegations" />
     </v-col>
     <v-col cols="12" md="4">
-      <ValidatorsViewDetailContentBottomRight />
+      <ValidatorsViewDetailDelegatorAmounts />
     </v-col>
   </v-row>
 </template>
 
 <script>
-import ValidatorsViewDetailContentBottomCenter from './ValidatorsViewDetailContentBottomCenter.vue';
-import ValidatorsViewDetailContentBottomLeft from './ValidatorsViewDetailContentBottomLeft.vue';
-import ValidatorsViewDetailContentBottomRight from './ValidatorsViewDetailContentBottomRight.vue';
+import ValidatorsViewDetailChart from './ValidatorsViewDetailChart.vue';
+import ValidatorsViewDetailDelegatedTokens from './ValidatorsViewDetailDelegatedTokens.vue';
+import ValidatorsViewDetailDelegatorAmounts from './ValidatorsViewDetailDelegatorAmounts.vue';
 
 import validatorsDetailDelegationsHelper from './helpers/validatorsDetailDelegationsHelper';
 import { mapGetters } from 'vuex';
@@ -27,9 +25,9 @@ import { mapGetters } from 'vuex';
 export default {
   name: 'ValidatorsViewDetailContentBottom',
   components: {
-    ValidatorsViewDetailContentBottomCenter,
-    ValidatorsViewDetailContentBottomLeft,
-    ValidatorsViewDetailContentBottomRight,
+    ValidatorsViewDetailChart,
+    ValidatorsViewDetailDelegatedTokens,
+    ValidatorsViewDetailDelegatorAmounts,
   },
   computed: {
     ...mapGetters('validators', ['account', 'delegations']),